+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [[0.2.3] - 2017-03-31]
 ### Added
-- Tests for empty cerberus response
+- Return error from Cerberus for auth calls
+- Tests for empty and error cerberus response
 
 ## [[0.2.2] - 2017-03-06]
 ### Fixed